final project/requirements.txt
Flask
psycopg2-binary
python-dotenv
Werkzeug